refactor(css): split site.css into base, topbar, site, cards modules (#67)
Original site.css (1731 lines) split into 4 files under css/shared/:
- base.css (373) - vars, fonts, reset, light theme, body styles
- topbar.css (796) - banner, nav, mobile drawer, search, footer
- site.css (429) - utilities, responsive nav, auth chip, legacy compat
- cards.css (183) - unified game card, toast

All 16 HTML pages updated with the 4 link tags. No CSS rules changed.
