+++ about.html
@@ -12,7 +12,10 @@
 <meta property="og:description" content="What Proton Pulse is, how it compares to ProtonDB, and how the open data pipeline works.">
 <meta property="og:type" content="website">
 <meta name="color-scheme" content="dark">
-<link rel="stylesheet" href="css/shared/site.css?v=aa1f96c6">
+<link rel="stylesheet" href="css/shared/base.css?v=45c7e06d">
+<link rel="stylesheet" href="css/shared/topbar.css?v=55e85c86">
+<link rel="stylesheet" href="css/shared/site.css?v=86052a10">
+<link rel="stylesheet" href="css/shared/cards.css?v=c0a5fad1">
 <link rel="stylesheet" href="css/index/index.css?v=0bdafa7e">
 </head>
 <body>
