+++ app.html
@@ -12,7 +12,10 @@
 <meta property="og:description" content="Browse merged ProtonDB community reports and hardware-scored Proton Pulse configs. Filter by GPU, driver, and Proton version. Searchable, filterable, and updated daily.">
 <meta property="og:type" content="website">
 <meta name="color-scheme" content="dark">
-<link rel="stylesheet" href="css/shared/site.css?v=aa1f96c6">
+<link rel="stylesheet" href="css/shared/base.css?v=45c7e06d">
+<link rel="stylesheet" href="css/shared/topbar.css?v=55e85c86">
+<link rel="stylesheet" href="css/shared/site.css?v=86052a10">
+<link rel="stylesheet" href="css/shared/cards.css?v=c0a5fad1">
 <link rel="stylesheet" href="css/app/app.css?v=c7bfd836">
 </head>
 <body>
@@ -50,6 +53,6 @@
 <script src="js/lib/topbar.js?v=38d23fe3"></script>
 <script src="js/lib/toast.js?v=7bb23014"></script>
 <script src="js/lib/gh-gist.js?v=346e3ea7"></script>
-<script type="module" src="js/app/main.js?v=89f78790"></script>
+<script type="module" src="js/app/main.js?v=681137a6"></script>
 </body>
 </html>
