+++ game-stats.html
@@ -9,7 +9,10 @@
 <title>Game Stats - Proton Pulse</title>
 <meta name="description" content="Per-game Proton compatibility analysis with confidence breakdown, monthly report trends, working status, version success rates, and proven launch options.">
 <meta name="color-scheme" content="dark">
-<link rel="stylesheet" href="css/shared/site.css?v=aa1f96c6">
+<link rel="stylesheet" href="css/shared/base.css?v=45c7e06d">
+<link rel="stylesheet" href="css/shared/topbar.css?v=55e85c86">
+<link rel="stylesheet" href="css/shared/site.css?v=86052a10">
+<link rel="stylesheet" href="css/shared/cards.css?v=c0a5fad1">
 <link rel="stylesheet" href="css/app/app.css?v=c7bfd836">
 <style>
 h1 {
