+++ profile.html
@@ -8,7 +8,10 @@
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <title>My Account — Proton Pulse</title>
 <meta name="color-scheme" content="dark">
-<link rel="stylesheet" href="css/shared/site.css?v=aa1f96c6">
+<link rel="stylesheet" href="css/shared/base.css?v=45c7e06d">
+<link rel="stylesheet" href="css/shared/topbar.css?v=55e85c86">
+<link rel="stylesheet" href="css/shared/site.css?v=86052a10">
+<link rel="stylesheet" href="css/shared/cards.css?v=c0a5fad1">
 <link rel="stylesheet" href="css/profile/profile.css?v=ebafcea9">
 </head>
 <body>
