+++ stats.html
@@ -9,7 +9,10 @@
 <title>Stats - Proton Pulse</title>
 <meta name="description" content="Aggregate compatibility stats for Steam on Linux: ratings, GPU and CPU breakdowns, OS distribution, Proton version share, and top-reported games. Filter by hardware to drill in.">
 <meta name="color-scheme" content="dark">
-<link rel="stylesheet" href="css/shared/site.css?v=aa1f96c6">
+<link rel="stylesheet" href="css/shared/base.css?v=45c7e06d">
+<link rel="stylesheet" href="css/shared/topbar.css?v=55e85c86">
+<link rel="stylesheet" href="css/shared/site.css?v=86052a10">
+<link rel="stylesheet" href="css/shared/cards.css?v=c0a5fad1">
 <style>
 h1 {
   font-family: var(--font-display);
